Move docker.go to its own sub-package
