+++ ecs-cli/modules/cli/local/docker/docker.go
@@ -11,13 +11,19 @@
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package local
+package docker
 
 import (
 	"os"
+	"time"
 
 	"github.com/docker/docker/client"
 	"github.com/sirupsen/logrus"
+)
+
+const (
+	// TimeoutInS is the wait duration for a response from the Docker daemon before returning an error to the user.
+	TimeoutInS = 30 * time.Second
 )
 
 const (
@@ -27,7 +33,8 @@ const (
 	minDockerAPIVersion = "1.27"
 )
 
-func newDockerClient() *client.Client {
+// NewDockerClient returns an object to communicate with the Docker Engine API.
+func NewDockerClient() *client.Client {
 	if os.Getenv("DOCKER_API_VERSION") == "" {
 		// If the user does not explicitly set the API version, then the SDK can choose
 		// an API version that's too new for the user's Docker engine.
